requirements.txt
streamlit
langflow
astrapy
python-dotenv
requests
openai